app completed

+++ backend/app.js
@@ -2,6 +2,7 @@ const express = require('express');
 const bodyParser = require('body-parser');
 const cors = require('cors');
 const { mongoose } = require('./db/mongoose')
+const jwt = require('jsonwebtoken');
 
 const app = express();
 
@@ -16,6 +17,22 @@ let corsOption = {
 
 app.use(cors(corsOption))
 app.use(bodyParser.json());
+
+// authentication for valid jwt token
+let authenticate = (req,res,next) => {
+    let token = req.header('x-access-token');
+
+    jwt.verify(token, User.getJWTSecret(), (err, decoded) => {
+        if(err) {
+            // JWT invalid
+            res.status(401).send(err);
+        } else {
+            // valid jwt
+            req.user_id = decoded._id
+            next();
+        }
+    });
+} 
 
 // verify refresh token middlewares
 let verifySession = (req, res, next) => {
@@ -65,19 +82,22 @@ app.get('/', (req,res) => {
 })
 
 // to get all the lists 
-app.get('/lists', (req,res) => {
-    List.find({}).then((lists) => {
+app.get('/lists', authenticate , (req,res) => {
+    List.find({
+        _userId: req.user_id
+    }).then((lists) => {
         res.send(lists);
     }).catch((err) => {
         res.send(err);
     });
 });
 
-// to create a list
-app.post('/lists', (req,res) => {
+// to create a list which belong to particular user authenticated user
+app.post('/lists', authenticate, (req,res) => {
     let title = req.body.title;
     let newList = new List({
-        title
+        title,
+        _userId: req.user_id
     });
     newList.save().then((listDoc) => {
         res.send(listDoc);
@@ -85,40 +105,61 @@ app.post('/lists', (req,res) => {
 });
 
 // update a spefic list
-app.patch('/lists/:id', (req,res) => {
-    List.findOneAndUpdate({ _id: req.params.id}, {
+app.patch('/lists/:id', authenticate, (req,res) => {
+    List.findOneAndUpdate({ _id: req.params.id, _userId: req.user_id}, {
         $set: req.body
     }).then(() => {
-        res.sendStatus(200)
+        res.send({'message': 'updated successfully!!'})
     })
 })
 
 // delete a list
-app.delete('/lists/:id', (req, res) => {
+app.delete('/lists/:id', authenticate, (req, res) => {
     List.findOneAndRemove({
-        _id: req.params.id
+        _id: req.params.id,
+        _userId: req.user_id
     }).then((removedListDoc) => {
-        res.send(removedListDoc)
+        res.send(removedListDoc);
+        // delete all task that are in the list
+        deleteTasksFromList(removedListDoc._id);
     })
 });
 
 // return all task to specific list
-app.get('/lists/:listId/tasks', (req,res) => {
+app.get('/lists/:listId/tasks', authenticate, (req, res) => {
     Task.find({
         _listId: req.params.listId
     }).then((tasks) => {
-        res.send(tasks)
+        res.send(tasks);
     })
-})
+});
 
 // creating a task
-app.post('/lists/:listId/tasks', (req,res) => {
-    let newTask = new Task({
-        title: req.body.title,
-        _listId: req.params.listId
-    })
-    newTask.save().then((newTaskDoc) => {
-        res.send(newTaskDoc);
+app.post('/lists/:listId/tasks', authenticate, (req,res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            // list object with the specified conditions was found
+            // therefore the currently authenticated user can create new tasks
+            return true;
+        }
+
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            let newTask = new Task({
+                title: req.body.title,
+                _listId: req.params.listId
+            });
+            newTask.save().then((newTaskDoc) => {
+                res.send(newTaskDoc);
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 })
 
@@ -133,24 +174,55 @@ app.post('/lists/:listId/tasks', (req,res) => {
 // })
 
 // updation of task
-app.patch('/lists/:listId/tasks/:taskId', (req,res) => {
-    Task.findOneAndUpdate({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }, {
-        $set: req.body
-    }).then(() => {
-        res.send({message: 'Updated successfully.'})
+app.patch('/lists/:listId/tasks/:taskId', authenticate, (req,res) => {
+
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            Task.findOneAndUpdate({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }, {
+                $set: req.body
+            }).then(() => {
+                res.send({'message': 'Updated successfully.'})
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 })
 
 // delete a task
-app.delete('/lists/:listId/tasks/:taskId', (req,res) => {
-    Task.findOneAndDelete({
-        _id: req.params.taskId,
-        _listId: req.params.listId
-    }).then((removedTaskDoc) => {
-        res.send(removedTaskDoc)
+app.delete('/lists/:listId/tasks/:taskId', authenticate ,(req,res) => {
+    List.findOne({
+        _id: req.params.listId,
+        _userId: req.user_id
+    }).then((list) => {
+        if (list) {
+            return true;
+        }
+
+        return false;
+    }).then((canCreateTask) => {
+        if (canCreateTask) {
+            Task.findOneAndDelete({
+                _id: req.params.taskId,
+                _listId: req.params.listId
+            }).then((removedTaskDoc) => {
+                res.send(removedTaskDoc)
+            })
+        } else {
+            res.sendStatus(404);
+        }
     })
 })
 
@@ -200,13 +272,22 @@ app.post('/users/login', (req, res) => {
 })
 
 // generate access token
-app.get("/users/me/access-token", verifySession, (req,res) => {
+app.get('/users/me/access-token', verifySession, (req, res) => {
     req.userObject.generateAccessAuthToken().then((accessToken) => {
-        res.header('x-access-token', accessToken).send({ accessToken })
-    }).catch((err) => {
-        res.status(400).send(err);
+        res.header('x-access-token', accessToken).send({ accessToken });
+    }).catch((e) => {
+        res.status(400).send(e);
+    });
+})
+
+// helper methods
+let deleteTasksFromList = (_listId) => {
+    Task.deleteMany({
+        _listId
+    }).then(() => {
+        console.log("done" + _listId)
     })
-});
+}
 
 app.listen(3000, () => {
     console.log("server is up and running!!!")
